refactor(fs): extract FS leaf math into shared pyo3-free fs-core (increment 1)
FS block scoring is the repo's parity orphan: it has THREE independent
implementations (Python numpy+scalar, the native pyo3 crate, and a hand-written
TS probabilistic.ts), while every other kernel follows the shared *-core ->
*-native + *-wasm triad. score-core owns only per-string similarity; the FS
math (banding/normalization/emit) is trapped in native/src/score.rs.

This starts making the Arrow-native kernel the single cross-surface source of
truth for FS. New standalone pyo3-free crate `goldenmatch-fs-core` (mirrors
score-core) now owns the two pure leaf functions `fs_normalize` (weight ->
[0,1], linear/posterior) and `fs_level_from_sim` (similarity -> comparison
level). score.rs re-exports them, so score_block_pairs_fs, the fused
match_fused_fs path, and the existing tests keep every call site unchanged --
parity holds by construction (same functions, relocated). Reference data is NOT
in the crate (design: injected by the host, never bundled).

Verified: fs-core cargo tests green; native rebuilds against fs-core; 249 FS
parity tests pass (test_probabilistic_vectorized / _bucket_native /
_native_authoritative / _native_fs_ne / nlevel_banding / _parallel /
_ne_scoring / _tf_scalar_parity); rustfmt clean.

Design note + full sequencing (move the block-scoring loop -> fs-core, then
fs-wasm + retire probabilistic.ts, then port the name/TF/ensemble scorers with
injected reference data, then delete the Python numpy/scalar production paths +
the FS_NATIVE/FS_VECTORIZED knobs + the numpy-only guard from the scale fix):
docs/superpowers/specs/2026-07-17-fs-core-cross-surface-extraction-design.md
