+++ packages/rust/extensions/fs-core/Cargo.toml
@@ -1,0 +1,23 @@
+# Standalone workspace so this pyo3-free Fellegi-Sunter core can be a path
+# dependency of the `native` crate (its own extension-module workspace) — and,
+# in later increments, of an `fs-wasm` binding — without either workspace
+# claiming it. Same isolation rationale as the sibling `score-core` /
+# `fingerprint-core` crates. NO rust-toolchain.toml here on purpose: the crate
+# inherits each parent crate's toolchain when built as a path dep.
+[workspace]
+
+[package]
+name = "goldenmatch-fs-core"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+authors = ["Ben Severn <benzsevern@gmail.com>"]
+description = "Canonical Fellegi-Sunter block-scoring math (level banding + weight normalization, no pyo3) shared across the native ext and the WASM/TS surface"
+
+[lib]
+name = "goldenmatch_fs_core"
+
+# Zero dependencies on purpose: this is pure f64 math. Per-string similarity
+# lives in `goldenmatch-score-core`; reference data (census / aliases) is
+# injected by the host, never bundled here (see the 2026-07-17 fs-core design).
+[dependencies]
